Improve admin maintenance boundaries
Split the large admin service implementation by functional responsibility while keeping public callers on the same AdminService surface. Move public admin behavior tests into integration tests and keep private auth-session state-machine coverage internal. Add repository guidance for English defaults in multilingual code paths.

Constraint: Keep admin split practical rather than overly granular.

Rejected: Exposing private auth-session helpers for integration tests | would widen API surface only for tests.

Confidence: high

Scope-risk: moderate

Directive: Keep multilingual fallback/default strings in English unless an explicit locale selects another language.

Tested: cargo fmt -p nyro-core; git diff --check; cargo test -p nyro-core --test admin; cargo test -p nyro-core admin::session_tests; cargo clippy -p nyro-core --all-targets

Not-tested: Full application UI smoke test.
